Moved global styles to utils and fixed spacing
